more tests, some refactor, removed "unitTests.isReturnDefaultValues = true" that I introduced for EventDetailsViewModel tests
