+++ app/src/test/java/com/github/se/eventradar/EventDetailsViewmodelUnitTest.kt
@@ -126,6 +126,4 @@ class EventDetailsViewmodelUnitTest {
     viewModel.getEventData()
     assert(!viewModel.isTicketFree())
   }
-
-
 }
